refactor(agent): extrae prompt base y bloques por-turno a agentPromptBase
Extracción VERBATIM (cero cambio de texto ni de orden) del system prompt
base, analyzeQuery, formatToolEvidence y los bloques de corpus/entidades/
análisis desde AgentScreen.jsx hacia un módulo puro. Habilita medir el
prompt ensamblado fuera de React (presupuesto de tokens en CI).
