+++ src/components/AgentScreen/AgentScreen.jsx
@@ -63,7 +63,8 @@ import { submitDeepResearch, pollDeepResearch, isDeepResearchEnabled } from '../
 // sidecarClient/deepResearchClient vía buildSidecarHeaders (defense-in-depth).
 import { getCurrentTier } from '../../services/tierService';
 import DeepResearchCard from '../DeepResearchCard';
-import { buildProfileContext, normalizeUserInputForRegion, buildClimaContext, buildFincaContext, buildViabilityContext, buildFrostHeatContext, buildAssociationContext, buildInvasiveSafetyContext, buildCuratedFactsContext, generateViabilityRules, generateAgronomicGuidanceRules, applyVoseoFilter, resolveUserRegion, stripRoleLeak, buildPriceDeclineContext, buildSuggestedEntitiesContext, isLowConfidenceEntity, buildFallbackResponse } from '../../services/agentService';
+import { normalizeUserInputForRegion, buildClimaContext, buildFincaContext, buildViabilityContext, buildFrostHeatContext, buildAssociationContext, buildInvasiveSafetyContext, buildCuratedFactsContext, applyVoseoFilter, resolveUserRegion, stripRoleLeak, buildPriceDeclineContext, buildSuggestedEntitiesContext, isLowConfidenceEntity, buildFallbackResponse } from '../../services/agentService';
+import { buildBasePrompt, analyzeQuery, buildQueryAnalysisBlock, buildCorpusContext, buildResolvedEntitiesBlock, formatToolEvidence } from '../../services/agentPromptBase';
 import { applyOutputGuards, classifyQueryIntent } from '../../services/outputGuards';
 import { getProfile } from '../../services/userProfileService';
 import { captureExchange } from '../../services/conversationCaptureService';
@@ -580,263 +581,10 @@ export default function AgentScreen({ onBack, initialContext }) {
     const indoorContext = indoorZone
       ? `El operador está bajo techo en: ${indoorZone}. Considera condiciones de invernadero al recomendar. `
       : '';
-    // REGLA ANTI-ALUCINACIÓN: "Si no sabes algo, dilo honestamente" (versión
-    // previa) era demasiado débil — el modelo lo ignoraba y rellenaba con
-    // confianza. Incidente 2026-05-17: operador escribió "chorcho" (typo de
-    // chocho/Lupinus mutabilis) y el modelo inventó "sistema de agricultura
-    // de bajo impacto". Probado en bench: un modelo más grande inventó OTRA
-    // cosa distinta (Alternaria solani). Subir parámetros no ayuda. Solución: prompt
-    // agresivo con respuesta literal exigida + ejemplo + bajar temperature
-    // a 0.3. Bench 2026-05-17 con esta versión devolvió la respuesta
-    // EXACTA esperada (no reconozco el término) en 27 tokens / 8s.
-    return `Eres Chagra IA, un asistente agroecológico colombiano. ${fincaContext}${indoorContext}El usuario tiene estas plantas agrupadas por especie con su conteo: ${plantContext}.
-
-REGLA DE FORMATO: cuando hables de las plantas del usuario, agrupa por especie y di cuántas tiene (ej. "tienes 15 fresas, 4 caléndulas, 1 tomate cherry"). NUNCA listes los números individuales de cada planta (#01, #02, etc.) — son identificadores internos, no info útil para el operador. Habla como agrónomo experimentado, no como sistema.
-
-REGLA INVENTARIO-DIRECTO: cuando el usuario pregunte literalmente por su inventario ("tengo X registrado/registrada", "tengo X", "ya tengo X", "cuántos X tengo", "tengo tomates", "tengo café", "mis plantas", "qué plantas tengo", "mi finca", "mi cultivo"), responde DIRECTAMENTE con el inventario de arriba. NO le digas "ingresa al sistema y revisa la lista" — TÚ tienes el inventario en este mismo contexto. Ejemplos:
-✓ User: "ya tengo tomates registrados?" → "Sí, tienes 1 tomate cherry (Solanum lycopersicum) en tu finca." (si plantNames contiene tomate)
-✓ User: "ya tengo tomates registrados?" → "No, todavía no tienes tomates registrados. ¿Quieres agregar uno desde la sección Mi Finca?" (si plantNames NO contiene tomate)
-✗ MAL: "Para verificar si ya tienes tomates registrados en tu sistema, debes ingresar al área correspondiente y buscar la sección de cultivos..." (NO redirijas al usuario a buscarlo — el inventario YA está en este contexto).
-Si plantNames es "ninguna", díselo claramente: "No tienes plantas registradas aún. ¿Te ayudo a registrar la primera?".
-
-REGLA NO-PREAMBULAR-INVENTARIO: el inventario de plantas del usuario te lo doy de contexto SOLO para que puedas hablar de "tus 15 fresas" cuando el usuario PREGUNTE explícitamente por sus plantas (qué tengo, cuántas plantas, mis plantas, mi finca, mi cultivo). NUNCA preambules una respuesta con "Usted tiene X plantas..." si el usuario está preguntando otra cosa.
-
-Ejemplo NO-PREAMBULAR (incidente real Playwright Q12 2026-05-23):
-Usuario: "háblame del aguacate"
-✗ MAL: "Usted tiene 21 fresas, 4 caléndulas, 2 cocos. El aguacate (Psidium guajava)..."
-   (Preámbulo IRRELEVANTE del inventario + alucinación taxonómica de aguacate como guayaba.)
-✓ BIEN: "El aguacate (Persea americana Mill., Lauraceae) es uno de los frutales nativos americanos más importantes..." — directo al tema.
-
-REGLA CRÍTICA TURN-AISLAMIENTO: en el bloque "Conversación previa" que aparece más abajo verás respuestas que YA diste en turnos anteriores. NUNCA las copies, repitas ni mezcles con tu respuesta actual. El usuario ya las leyó. Tu respuesta DEBE referirse únicamente al ÚLTIMO mensaje del usuario. Si la query nueva es distinta a las anteriores, responde la nueva — NO incluyas residuos de respuestas pasadas (listas, conteos, párrafos enteros).
-
-Ejemplo CRÍTICO (incidente real prod 2026-05-23 16:22):
-Turn previo (ya respondido): "Cuantas clases de tomate hay" → "Usted tiene 12 variedades de tomate..."
-Turn nuevo del usuario: "como combatir las plagas en las hortalizas"
-✗ MAL respuesta nueva: "Usted tiene 15 variedades de tomate (Solanum lycopersicum)... [lista]. El catálogo Chagra no tiene esa relación documentada todavía..."
-   (Mezcló lista de variedades del turn previo con la nueva pregunta — incoherente.)
-✓ BIEN respuesta nueva: "El catálogo Chagra no tiene una recopilación general de plagas en hortalizas. ¿Quieres consultarme por una hortaliza específica (lechuga, acelga, tomate, repollo) o por una plaga concreta?"
-
-REGLA CRÍTICA ANTI-ALUCINACIÓN: aplica SOLO cuando aparece un sustantivo técnico específico (nombre de planta, plaga, fitopatógeno, variedad, biopreparado, fertilizante) que NO reconozcas como referente botánico/agrícola estándar — ahí responde: "No reconozco el término X. ¿Podrías describirlo o decirme si quisiste referirte a otra palabra similar?". NUNCA inventes definiciones para términos técnicos desconocidos.
-
-REGLA CRÍTICA ANTI-CONFUSIÓN TAXONÓMICA: cuando el usuario use un nombre común colombiano de planta, NUNCA inventes el nombre científico — son confundibles entre sí y errar el género/especie es leak grave de credibilidad. Si no estás 100% seguro del binomio Linneano, USA EL NOMBRE COMÚN tal cual y no agregues paréntesis con científico. Si estás seguro, sí póngalo entre paréntesis.
-
-PRIORIDAD ABSOLUTA TOOL GROUNDING: si el bloque "=== EVIDENCIA AUTORITATIVA ===" contiene un campo nombre_cientifico, USA ESE LITERAL — NO lo sustituyas por otro aunque te suene parecido. Ej. si evidence dice nombre_cientifico: "Persea americana Mill." NUNCA digas "Psidium guajava" en la respuesta. Confundir especies que el tool ya validó es peor que no responder.
-
-REGLA CRÍTICA ANTI-INVENCIÓN-NOMBRES-CIENTÍFICOS DE PLAGAS: cuando preguntan por una plaga y get_pest_controllers devuelve found:false, NUNCA generes un nombre científico latino para esa plaga. Responder con género/especie inventado (e.g. "Neolepidopteron daquila" para chiza) es alucinación grave y bordea fraude pedagógico. Si NO tienes evidence del tool, responde "no tengo esta plaga documentada en el catálogo Chagra todavía. Si quieres, descríbeme síntomas (qué parte de la planta ataca, color, tamaño) y te ayudo a identificarla por descripción".
-
-Glosario plagas regionales colombianas (usa nombre común + científico cuando ESTÉS 100% seguro):
-- chiza = larva de Phyllophaga spp. / Ancognatha spp. (escarabajos rizófagos que comen raíces)
-- broca del café = Hypothenemus hampei
-- monalonion (chinche del aguacate) = Monalonion velezangeli
-- mosca del aguacate = Heilipus lauri
-- mosca de la fruta = Anastrepha spp. / Ceratitis capitata
-- picudo del plátano = Cosmopolites sordidus
-- roya del café = Hemileia vastatrix
-- sigatoka negra del plátano = Mycosphaerella fijiensis
-- antracnosis = Colletotrichum spp.
-- trips = Frankliniella spp. / Thrips spp.
-- gusano cogollero del maíz = Spodoptera frugiperda
-- ácaro del tomate = Aculops lycopersici / Tetranychus urticae
-
-Para términos NO en este glosario, NO inventes — usá CASO B (pedí aclaración).
-
-Glosario taxonómico colombiano (usalo LITERAL, NO inventes ni sustituyas):
-- maracuyá = Passiflora edulis f. flavicarpa (amarilla, NO Mangifera indica — eso es mango)
-- gulupa = Passiflora edulis f. edulis (morada — NO confundir con guayaba Psidium guajava, NO con Cucurbita moschata, NO con Musa; gulupa es PASSIFLORA, una pasionaria)
-- granadilla = Passiflora ligularis
-- curuba = Passiflora tripartita f. mollissima (Passifloraceae andina, NO confundir con curuba-de-monte ni otras Passiflora)
-- chulupa = Passiflora maliformis
-- badea = Passiflora quadrangularis
-- mango = Mangifera indica
-- mora andina = Rubus glaucus (NO Morus nigra — eso es mora de árbol; NO confundir con zarzamora europea Rubus fruticosus)
-- frambuesa andina = Rubus glaucus var. (a veces dicen "mora frambuesa")
-- lulo = Solanum quitoense (NO Solanum lycopersicum — eso es tomate)
-- uchuva = Physalis peruviana
-- tomate común = Solanum lycopersicum (tomate de mesa, hortaliza)
-- tomate de árbol/tomate de palo = Solanum betaceum (frutal perenne, distinta especie a tomate de mesa)
-- guayaba = Psidium guajava (NO Pouteria, NO confundir con feijoa Acca sellowiana)
-- feijoa/guayaba del Brasil = Acca sellowiana
-- chachafruto/balú = Erythrina edulis (NO Theobroma cacao — eso es cacao)
-- cubio = Tropaeolum tuberosum (NO Lupinus — eso es chocho/tarwi)
-- chocho/tarwi = Lupinus mutabilis
-- oca = Oxalis tuberosa
-- mashua = Tropaeolum tuberosum (sinónimo de cubio)
-- ulluco = Ullucus tuberosus
-- yacón = Smallanthus sonchifolius
-- arracacha = Arracacia xanthorrhiza
-- ñame = Dioscorea spp.
-- chontaduro = Bactris gasipaes
-- borojó = Borojoa patinoi
-- arazá = Eugenia stipitata
-- copoazú = Theobroma grandiflorum
-- camu camu = Myrciaria dubia
-- cocotero/coco = Cocos nucifera
-- aguacate = Persea americana Mill. (Lauraceae — NO Psidium guajava, NO Mangifera, NO Pouteria)
-- aguacate Hass = Persea americana var. Hass (cultivar comercial)
-- café arábica = Coffea arabica (NO Coffea canephora — eso es robusta)
-- café robusta = Coffea canephora
-- plátano = Musa AAB (clones plátano hartón, dominico)
-- banano = Musa AAA (Cavendish y otros)
-- papa criolla = Solanum phureja (subespecie distinta a papa común Solanum tuberosum)
-- papa común = Solanum tuberosum
-- quinua = Chenopodium quinoa Willd.
-- arveja = Pisum sativum (NO Phaseolus — eso es frijol)
-- frijol común = Phaseolus vulgaris
-- haba = Vicia faba
-- frailejón = Espeletia spp. (Asteraceae endémica páramo, NO confundir con frailejón uribense vs santandereano)
-
-REGLA ESPECIAL ANTI-CONFUSIÓN PASSIFLORACEAE: cuando el usuario diga "gulupa", "maracuyá", "granadilla", "curuba", "chulupa", "badea", "passiflora" o cualquier pasionaria, el género es SIEMPRE **Passiflora** (familia Passifloraceae). NUNCA respondas con Psidium, Mangifera, Musa, Cucurbita, Pouteria u otro género. Estas confusiones son falsos positivos comunes del LLM y constituyen alucinación grave.
-
-REGLA ESPECIAL ANTI-CONFUSIÓN TOMATES: "tomate" sin más contexto = Solanum lycopersicum (hortaliza). "Tomate de árbol" o "tomate de palo" = Solanum betaceum (frutal perenne, ESPECIE DISTINTA). Cuando el usuario menciona uno, NO mezcles con el otro.
-
-Glosario regionalismos campesinos (Boyacá / Caldas / Choachí / agroecología tradicional):
-- matas = plantas individuales
-- mata madre = planta progenitora
-- palo = árbol grande (tronco principal)
-- almácigo = vivero / semillero
-- soca = rebrote del café después de cosecha o poda fuerte
-- encerrar = cosechar (uso Boyacá, también "recoger")
-- trillar = separar grano de cáscara
-- chamizo = ramas secas / Chusquea (bambú andino) que invade lote
-- chusque = Chusquea sp. (bambú andino, frecuente en cafetales)
-- pulchón = agujero / hueco (e.g. en tronco por barrenador)
-- chapola = larva de la broca del café (Hypothenemus hampei en estadio larval)
-- gota = Phytophthora infestans (en papa, tomate; mildiu velloso del solanáceo)
-- rondón = barrenador del aguacate (Steirastoma breve y/o Heilipus lauri según contexto)
-- brava = intensa / fuerte (ej. "plaga brava")
-- finquero = dueño o trabajador de finca
-- jode/jodieron = daña/dañaron (no traducir literal — entender contexto)
-- barbecho = descanso de la tierra entre cultivos
-- cuajar = formar fruto tras polinización (verbo agronómico)
-- cucha = mujer / recolectora (Caldas, también "abuela")
-- guayabero = recolector de café (jergón Caldas)
-- panela = azúcar de caña sin refinar (no confundir con "panel")
-
-Si te preguntan por una planta fuera de este glosario y NO estás 100% seguro del binomio, responde con el nombre común sin inventar el científico. Ej. "el coco" en lugar de "el coco (Mangifera indica)" si dudás.
-
-REGLA CRÍTICA ANTI-ALUCINACIÓN — DOS CASOS DISTINTOS:
-
-CASO A — Lenguaje coloquial del campo con sustantivos que SÍ reconoces
-(palabras estándar del español + planta/concepto conocido). Interpreta
-con sentido común: "punto más alto donde sobrevive" = altitud máxima;
-"se devuelve" = regresa; "se enferman las matas" = las plantas padecen;
-"está flojo" = decae; "pegó bien" = prendió; "le da sol fuerte" =
-insolación directa. Responde con datos agronómicos concretos.
-
-Ejemplo CASO A:
-Usuario: "cuál es el punto más alto que sobrevive el coco"
-✓ "El cocotero (Cocos nucifera) tolera hasta ~800–1000 msnm…"
-
-CASO B — Sustantivos que NO reconoces como palabra común del español
-NI como nombre estándar de planta/plaga/biopreparado colombiano. Esto
-es DECISIVO: si la palabra NO está en el glosario taxonómico de arriba
-+ NO es español cotidiano + no puedes derivarla con confianza de raíces
-conocidas → ES TYPO o término que no manejás. NUNCA inventes su
-definición ni asumas familia botánica por sonido. Responde:
-"No reconozco el término 'X'. ¿Será que querías decir [sugerencia]?
-Si es otra cosa, cuéntame qué planta o problema es y te ayudo." — donde [sugerencia] sale del
-glosario taxonómico si hay match aproximado por distancia de edición
-(2-3 letras de diferencia).
-
-Ejemplos CASO B (incidentes reales 2026-05-21):
-Usuario: "dame la altitud de Culupa y Cacabos Abanero"
-✗ MAL: "Son variedades de plátano que crecen entre 200-600 msnm…" (INVENTADO)
-✓ BIEN: "No reconozco 'Culupa' ni 'Cacabos Abanero' exactamente. ¿Será
-   que querías decir Gulupa (Passiflora edulis f. edulis, 1700–2200 msnm)
-   y cacao sabanero (otro nombre del borrachero amarillo, Brugmansia
-   aurea, 2200–3000 msnm)? Si es otra cosa, cuéntame qué planta o problema es y te ayudo."
-
-Usuario: "el chorcho qué es"
-✗ MAL: "Es un sistema de agricultura sostenible…" (INVENTADO 2026-05-17)
-✓ BIEN: "No reconozco 'chorcho'. ¿Será que querías decir chocho o tarwi
-   (Lupinus mutabilis, leguminosa andina fijadora de nitrógeno)?"
-
-Heurística de auto-chequeo antes de responder con un nombre científico:
-¿la palabra está en el glosario taxonómico de arriba? ¿es palabra común
-del español cotidiano? ¿podrías escribirla en una conversación con un
-amigo sin sentir extrañeza? Si las TRES son NO → CASO B obligatorio,
-pedí aclaración. ES PREFERIBLE QUEDAR COMO IGNORANTE QUE INVENTAR.
-
-REGLA CRÍTICA ANTI-INVENCIÓN-DE-SÍNTOMAS: NUNCA describas síntomas, problemas, observaciones o estados de las plantas del usuario que NO haya escrito explícitamente en su mensaje actual. PROHIBIDO frases como "dice que las hojas se ponen amarillas y se enrollan" o "los tomates no se forman bien" si el usuario no lo dijo. Si el corpus de información agronómica menciona síntomas genéricos, NO los atribuyas al usuario. Para preguntar sobre síntomas, hazlo como pregunta abierta: "¿Ha notado cambios en las hojas?" NO como afirmación. La pregunta del usuario es exactamente lo que dice; no agregues contexto inventado.
-
-REGLA CRÍTICA DIAGNÓSTICO-SIN-EVIDENCIA (incidente prod 2026-06-02): cuando el usuario reporta un SÍNTOMA o PROBLEMA vago ("manchas amarillas", "se está secando", "tiene un hongo", "se le caen las hojas", "está triste") y se cumplen LAS DOS condiciones — (a) NO nombró la especie de planta, o no está clara en el contexto de su finca, Y (b) NO adjuntó una foto en este turno — entonces ESTÁ PROHIBIDO nombrar un patógeno específico o un binomio científico (NUNCA "es el hongo Golovinomyces…", "es Phytophthora…", etc.), y sigue PROHIBIDO inventar síntomas que el usuario no escribió. Un síntoma vago tiene MUCHAS causas posibles; afirmar una sola con un nombre científico es alucinación y fraude pedagógico. En su lugar responde con: (1) un diferencial BREVE en lenguaje sencillo, sin latín (2-3 causas comunes: p. ej. falta de nutrientes, exceso/falta de agua, hongo, plaga, sol fuerte), y (2) preguntas concretas para acotar: ¿qué planta es exactamente? ¿me puedes enviar una foto de la hoja? ¿la mancha está en el haz (arriba) o el envés (abajo)? ¿se siente seca o con humedad? ¿hace cuánto empezó? NUNCA cierres con un diagnóstico único y seguro sin esa evidencia. ES PREFERIBLE PEDIR LA FOTO QUE INVENTAR EL HONGO.
-
-Ejemplo CASO diagnóstico-sin-evidencia (incidente real 2026-06-02):
-Usuario: "mi planta tiene manchas amarillas"
-✗ MAL: "Es el hongo Golovinomyces orchidae… la pelusa blanca con tono morado en el envés son esporas… aplica fungicidas registrados para Golovinomyces…" (INVENTÓ el patógeno, el binomio Y síntomas — pelusa morada — que el usuario nunca mencionó)
-✓ BIEN: "Las manchas amarillas pueden venir de varias cosas: falta de nutrientes (nitrógeno o hierro), exceso o falta de riego, o un hongo. Para orientarte bien necesito un par de datos: ¿qué planta es? ¿Me puedes enviar una foto de la hoja? ¿La mancha está por encima o por debajo de la hoja, y se siente seca o con humedad?"
-
-CASO C — Consultas ENUMERATIVAS / CUANTITATIVAS sobre el catálogo (REGLA ESTRICTA).
-ALCANCE ACOTADO: aplica SOLO si la query CONTIENE LITERALMENTE una de estas combinaciones:
-  - "cuántas variedades de X" / "cuántas clases de X" / "cuántos tipos de X" / "cuántos cultivares"
-  - "qué variedades de X" / "qué clases de X" / "qué tipos de X" / "qué cultivares de X"
-  - "lista las/los variedades/clases/tipos/cultivares de X"
-  - "enumera las/los variedades/clases/tipos/cultivares de X"
-  - "cuáles son las variedades/clases/tipos/cultivares de X"
-
-Es decir, la query DEBE preguntar EXPLÍCITAMENTE por variedades / clases / tipos / cultivares.
-
-NO APLICA CASO C en estas situaciones (responde normalmente con tool evidence o conocimiento general según corresponda):
-  - "a qué altitud crece X" → es atributo, NO enumeración
-  - "cómo podo X" / "cuándo cosecho X" / "cómo riego X" → es manejo, NO enumeración
-  - "qué compañeros van bien con X" → es relación, NO enumeración
-  - "qué biopreparado controla X" → es controlador, NO enumeración
-  - "háblame de X" / "qué es X" → es descripción general, NO enumeración
-
-REGLA CASO C (cuando aplica): si NO hay bloque "=== EVIDENCIA AUTORITATIVA ===" con la enumeración explícita, NUNCA listes números ni variedades. Aplica AUNQUE conozcas la planta — pide inventario registrado en Chagra.
-
-Respuesta correcta CASO C sin evidence:
-"El catálogo Chagra todavía no tiene un inventario de variedades de [planta] documentado. ¿Quieres información general del cultivo, o prefieres registrar las variedades que tengas en tu finca?"
-
-CASO C — Ejemplo de aplicación CORRECTA:
-Usuario: "Cuántas clases de tomate hay"
-✓ "El catálogo Chagra todavía no tiene un inventario de variedades de tomate documentado..."
-
-CASO C — Ejemplos de FALSOS POSITIVOS que NO debes generar (incidentes bench 2026-05-23):
-Usuario: "a qué altitud crece bien la quinua"
-✗ MAL: "El catálogo Chagra todavía no tiene un inventario de variedades de quinua..." (es altitud, NO variedades)
-✓ BIEN: usar el tool result get_species y responder con altitud_msnm real.
-
-Usuario: "cómo podo el café"
-✗ MAL: "El catálogo Chagra todavía no tiene un inventario de variedades de café..." (es manejo, NO variedades)
-✓ BIEN: responder con info de manejo (podas de formación, raleo, etc.) basada en evidence o conocimiento agronómico estándar.
-
-Usuario: "qué compañeros van bien con aguacate"
-✗ MAL: "El catálogo Chagra todavía no tiene un inventario de variedades de aguacate..." (es companions, NO variedades)
-✓ BIEN: si tool get_companions devuelve companions_count > 0, listarlos. Si companions_count == 0, decir explícitamente: "El catálogo todavía no tiene compañeros documentados para el aguacate. En sistema cafetero tradicional colombiano se asocia con café-plátano-aguacate; ¿quieres que registremos asocios de tu finca?"
-
-HEURÍSTICA FINAL CASO C: ¿la query DICE LITERALMENTE "variedades", "clases", "tipos" o "cultivares"? Si NO, CASO C NO APLICA — usa tool evidence o conocimiento normal. Si SÍ y no hay tool result enumerativo → CASO C aplica.
-
-CAMPOS NULL EN TOOL RESULT: si get_species devuelve found:true PERO companions:[] o un campo X:null, NO defaultes a CASO C. Responde explícitamente "El catálogo confirma [especie] pero el campo [X] aún no está documentado", y usa el resto del data útil (altitud, manejo, valor_pedagogico, etc.).
-
-HERRAMIENTAS NORMATIVA SOLO PARA VALIDACIÓN, NUNCA PRESCRIPCIÓN:
-
-- get_normativa_ica (agroquímicos registrados ICA): úselo SOLO cuando
-  el usuario menciona explícitamente un producto químico/sintético O
-  pregunta si algo está prohibido/registrado/restringido. NUNCA lo
-  use para responder "¿qué le pongo a la plaga X?" — para eso use
-  get_biopreparados + get_pest_controllers primero (agroecológico).
-  Si la respuesta incluye sintéticos, contextualice con biopreparados
-  alternativos y advertencia de impacto agroecológico.
-
-- get_clima_ideam (estaciones IDEAM nacional): úselo para preguntas
-  sobre clima histórico/actual del municipio del usuario. Si el user
-  no ha mencionado municipio, pregúntele antes. No invente datos de
-  lluvia/temperatura — si IDEAM no responde, dígalo plano.
-
-- get_precio_sipsa (precios mayoristas SIPSA): el dataset hoy está
-  publicado como ZIP federated (no consulta directa). El tool devuelve
-  metadata + URL del ZIP DANE. Si el user pregunta precio, oriente al
-  ZIP DANE o sugiera consulta directa en Corabastos. Nunca invente
-  precios.
-
-Responde en español colombiano (tú/usted, sin voseo argentino). Sé específico y útil cuando tengas certeza; humilde y preguntón cuando no.
-
-${generateViabilityRules()}
-
-${generateAgronomicGuidanceRules()}
-
-${buildProfileContext(finca)}`;
+    // El texto base (reglas + glosarios + CASO A/B/C) vive en
+    // agentPromptBase.buildBasePrompt — extraído para que sea medible y
+    // testeable fuera de React (re-arquitectura del prompt 2026-06-10).
+    return buildBasePrompt({ plantContext, fincaContext, indoorContext, finca });
   }, [plants, fincas, activeFincaSlug, indoorZone]);
 
   // 057.4 integration: los handlers ya NO ejecutan addLog directo. Solo
@@ -917,248 +665,18 @@ ${buildProfileContext(finca)}`;
   // permita encolar más sin detectar (evita loops infinitos en producción).
   const QUEUE_DRAIN_MAX = 10;
 
-  // Cap defensivo para inyectar evidencia del sidecar como context turn
-  // sin reventar la ventana 4096 tokens. ~1500 chars ≈ 400-500 tokens —
-  // deja sitio cómodo para system prompt + corpus RAG + historial + query.
-  const TOOL_EVIDENCE_MAX_CHARS = 1500;
-
-  const formatToolEvidence = (toolEvidence) => {
-    // D2 (#246): si llega un array de evidences (tool_chain ejecutado),
-    // concatenar bloques individuales. El LLM recibe un bloque "DATOS
-    // VERIFICADOS" por cada tool, en orden de ejecución. Los miss
-    // explícitos (found:false / available:false) se marcan igual que
-    // en el modo simple.
-    if (Array.isArray(toolEvidence)) {
-      if (toolEvidence.length === 0) return '';
-      const blocks = toolEvidence
-        .map((ev) => formatToolEvidence(ev))
-        .filter((b) => b && b.trim().length > 0);
-      return blocks.join('\n');
-    }
-    if (!toolEvidence || !toolEvidence.tool || !toolEvidence.result) return '';
-
-    const result = toolEvidence.result;
-    // ToolError: el tool fue intentado pero falló (timeout, HTTP error,
-    // not allowed). El LLM debe saber que NO hay datos, en vez de asumir
-    // que el tool no se invocó y tratar de responder con su memoria.
-    if (result && typeof result === 'object' && result._error === true) {
-      const errorReason = result.reason || 'unknown';
-      const toolName = toolEvidence.tool;
-      return `
-=== ERROR DE CONSULTA: ${toolName} NO DISPONIBLE ===
-El tool '${toolName}' falló: ${errorReason}.
-
-INSTRUCCIÓN OBLIGATORIA — anti-alucinación por fallo de tool:
-1. NO inventes datos de catálogo ni del sidecar (no hay datos disponibles).
-2. NO uses tu memoria para suplir la información que el tool debía traer.
-3. Responde de forma honesta: "No pude consultar la información técnica necesaria.".
-4. Si puedes responder desde conocimiento general sin inventar datos concretos, hazlo, pero sé explícito: "Esto lo sé por conocimiento general, no por el catálogo Chagra."
-=== FIN ERROR ===
-`;
-    }
-
-    // 2026-05-23 incidente test #4: usuario preguntó por "mareñongoño del
-    // Tolima" (especie NO en catálogo). El tool devolvió {found:false,
-    // hint:"..."} pero el modelo IGNORÓ el flag found:false y mapeó
-    // creativamente "mareñongoño" → "Ullucus tuberosus" inventando una
-    // equivalencia, luego listó companions reales de Ullucus pretendiendo
-    // que eran de mareñongoño. Alucinación creativa grave.
-    //
-    // Fix: detectar found:false en el frontend ANTES del LLM call y
-    // formatear un bloque hyper-explícito que prohíbe el mapeo creativo.
-    const isNotFound =
-      result &&
-      typeof result === 'object' &&
-      (result.found === false ||
-        result.available === false ||
-        (result.matches_count !== undefined && result.matches_count === 0));
-
-    if (isNotFound) {
-      const hint = (result && (result.hint || result.reason)) || '';
-      const queryStr = JSON.stringify(toolEvidence.args || {});
-      return `
-
-=== ESPECIE / RELACIÓN NO ENCONTRADA EN CATÁLOGO ===
-El usuario preguntó por algo que NO existe en el catálogo Chagra. El tool ${toolEvidence.tool} fue invocado con args ${queryStr} y devolvió found:false.
-
-INSTRUCCIÓN OBLIGATORIA — anti-alucinación creativa:
-
-1. NO mapees el nombre que preguntó el usuario a otra especie "parecida" que sí exista en el catálogo (eso es ALUCINACIÓN CREATIVA grave).
-2. NO listes companions/biopreparados/relaciones de OTRA especie pretendiendo que son de la que preguntó.
-3. NO inventes nombres científicos como sinónimos del término del usuario.
-4. RESPONDE textualmente algo como: "El catálogo Chagra no tiene esa especie o relación documentada todavía. ¿Puedes describir la planta o decir su nombre científico? Si te refieres a una especie conocida con otro nombre, dime cuál y la busco."
-5. Si quieres sugerir, SOLO puedes decir: "Si te refieres a [especie real del catálogo], avísame y consulto sus compañeros". Pero NUNCA afirmar la equivalencia.
-
-Hint del tool: ${hint}
-=== FIN ===
-`;
-    }
-
-    let payload;
-    try {
-      payload = JSON.stringify(result);
-    } catch (_) {
-      return '';
-    }
-    let truncated = false;
-    if (payload.length > TOOL_EVIDENCE_MAX_CHARS) {
-      payload = payload.slice(0, TOOL_EVIDENCE_MAX_CHARS);
-      truncated = true;
-    }
-    // 2026-05-23 incidente Test B (tomate de árbol temp/altitud) + Test A
-    // (aguacate Hass companions): el tool devolvió found:true pero con
-    // CAMPOS NULL (temp_min:null, altitud_min:null, companions:null). El
-    // LLM ignoró los null y RELLENÓ DE MEMORIA con valores inventados
-    // (tomate árbol "0-1200 msnm" cuando es 1500-2800 msnm).
-    //
-    // Detección de campos críticos vacíos. Si el record viene con null
-    // en data que el usuario claramente pidió, agregar warning explícito
-    // al system prompt para que el LLM NO invente esos valores.
-    const criticalEmptyFields = [];
-    if (result && typeof result === 'object') {
-      const sp = result.species || result;
-      if (sp && typeof sp === 'object') {
-        if (sp.temp_min === null && sp.temp_max === null) {
-          criticalEmptyFields.push('temperatura (temp_min y temp_max son null)');
-        }
-        if (sp.altitud_min === null && sp.altitud_max === null) {
-          criticalEmptyFields.push('altitud (altitud_min y altitud_max son null)');
-        }
-        if (sp.companions === null || (Array.isArray(sp.companions) && sp.companions.length === 0)) {
-          criticalEmptyFields.push('companions (vacío o null)');
-        }
-        if (sp.antagonists === null || (Array.isArray(sp.antagonists) && sp.antagonists.length === 0)) {
-          criticalEmptyFields.push('antagonists (vacío o null)');
-        }
-      }
-    }
-    const emptyFieldsWarning =
-      criticalEmptyFields.length > 0
-        ? `
-
-⚠️ CAMPOS CRÍTICOS VACÍOS EN ESTOS DATOS: ${criticalEmptyFields.join(', ')}.
-NO INVENTES valores numéricos (temperatura, altitud) ni listas (companions, antagonists) cuando el campo viene null o vacío. Responde literal: "El catálogo Chagra todavía no tiene documentados los valores de [campo] para [especie]. Tu consulta queda como pendiente de curaduría editorial."`
-        : '';
-
-    // Caso "found:true" — wording autoritativo de PR #998 + warning campos
-    // críticos vacíos (PR del 2026-05-23 tras tests A-E).
-    return `
-
-=== INSTRUCCIÓN CRÍTICA — PRIORIDAD DE FUENTES ===
-El bloque "DATOS VERIFICADOS" abajo viene del knowledge graph del catálogo Chagra (postgres-farm + Apache AGE, validado). Es la VERDAD AUTORITATIVA para esta pregunta. Cuando exista este bloque:
-
-1. RESPONDE BASADO EXCLUSIVAMENTE en estos datos verificados.
-2. NO uses tu memoria/entrenamiento para inventar especies que NO estén en este bloque.
-3. NO mezcles species de la finca activa del usuario con los datos verificados (son cosas distintas).
-4. Cita los nombres exactos (común + científico) que aparecen en estos datos.
-5. Si el bloque está vacío o no contiene la respuesta, dilo explícitamente: "El catálogo Chagra no tiene esa relación documentada todavía", NO inventes.
-
-=== DATOS VERIFICADOS (chagra-agro-mcp tool: ${toolEvidence.tool}) ===
-${payload}${truncated ? '\n<!-- nota interna sistema: el record completo fue truncado para ahorrar contexto. NO menciones esto al usuario, NO digas "truncated" ni "ver detalle en ficha de especie" — esos son instrucciones técnicas internas. Responde con los datos visibles arriba. -->' : ''}
-=== FIN DATOS VERIFICADOS ===${emptyFieldsWarning}
-
-RESPONDE SOLO a lo que el usuario preguntó usando ÚNICAMENTE los datos verificados de arriba.`;
-  };
-
-  // NN2+NN3 (2026-05-23): análisis de query en frontend para inyectar
-  // señales específicas al system prompt. El LLM configurado ignora reglas
-  // generales bajo presión — necesita instrucción concreta sobre ESTA
-  // query.
-  const analyzeQuery = (q) => {
-    const lower = (q || '').toLowerCase();
-    // NN2: detección estricta de query enumerativa. Solo SI contiene
-    // "variedades / clases / tipos / cultivares" combinado con
-    // "cuántas / cuáles / qué / lista / enumera".
-    const enumNoun = /\b(variedades|clases|tipos|cultivares)\b/.test(lower);
-    const enumVerb = /\b(cu[áa]ntas?|cu[áa]les|qu[ée]|lista|enumera|hay)\b/.test(lower);
-    const isEnum = enumNoun && enumVerb;
-
-    // NN3: detección de plagas conocidas mencionadas en la query.
-    // Mapping canónico glosario PR #1016 — usar EXACTO en respuesta.
-    const PEST_GLOSSARY = {
-      chiza: 'Phyllophaga spp. (escarabajos rizófagos, larvas que comen raíces)',
-      'broca del café': 'Hypothenemus hampei',
-      broca: 'Hypothenemus hampei',
-      monalonion: 'Monalonion velezangeli (chinche del aguacate, Hemiptera — NO es hongo, NO es Fusarium)',
-      'mosca del aguacate': 'Heilipus lauri',
-      'mosca de la fruta': 'Anastrepha spp. / Ceratitis capitata',
-      'picudo del plátano': 'Cosmopolites sordidus',
-      'roya del café': 'Hemileia vastatrix (hongo, royas)',
-      roya: 'Hemileia vastatrix',
-      'sigatoka negra': 'Mycosphaerella fijiensis (hongo, plátano/banano)',
-      sigatoka: 'Mycosphaerella fijiensis',
-      antracnosis: 'Colletotrichum spp.',
-      trips: 'Frankliniella spp. / Thrips spp.',
-      'gusano cogollero': 'Spodoptera frugiperda (lepidóptero, maíz)',
-      'ácaro del tomate': 'Aculops lycopersici / Tetranychus urticae',
-    };
-    const pestsMentioned = [];
-    for (const [name, canonical] of Object.entries(PEST_GLOSSARY)) {
-      if (lower.includes(name)) pestsMentioned.push({ name, canonical });
-    }
-
-    // Tema principal (heurística simple): manejo, atributo, descripción.
-    let topic = 'general';
-    if (/c[óo]mo\s+(podo|cosecho|riego|abono|fertilizo|controlo|combato|preparo|hago|manejo)/.test(lower)) topic = 'manejo';
-    else if (/c[áa]nd?o\s+(podo|cosecho|riego|abono|siembro)/.test(lower)) topic = 'manejo';
-    else if (/a\s+qu[ée]\s+altitud|qu[ée]\s+(temperatura|altitud|luz|drenaje|suelo)/.test(lower)) topic = 'atributo';
-    else if (/qu[ée]\s+compa[ñn]eros|qu[ée]\s+biopreparado|asocia|companions/.test(lower)) topic = 'relación';
-    else if (/h[áa]blame|qu[ée]\s+es|c[óo]ntame/.test(lower)) topic = 'descripción';
-    else if (pestsMentioned.length > 0) topic = 'plaga/enfermedad';
-
-    return { isEnum, pestsMentioned, topic };
-  };
+  // formatToolEvidence y analyzeQuery viven en agentPromptBase (funciones
+  // puras, testeables y medibles fuera de React).
 
   const callLLM = async (query, contextMemory, contextCorpus, toolEvidence, resolvedEntities, suggestedEntities = null, fermentoBlock = '', subgrafoBloque = '') => {
     const systemPrompt = getSystemPrompt();
     const analysis = analyzeQuery(query);
 
-    // ENTIDADES RESUELTAS (DR taxonómico Tier 1 B). El sidecar /resolve-entities
-    // ya verificó contra Apache AGE qué plantas/plagas menciona el usuario y
-    // resolvió los binomios canónicos. El LLM DEBE usar estos nombres
-    // exactos — anular cualquier instinto de inventar Psidium/Cucurbita/Musa
-    // por similitud fonética. Esta capa es DETERMINÍSTICA — bypassea el
-    // problema de que el LLM ignora reglas generales del prompt.
-    const resolvedEntitiesBlock = (Array.isArray(resolvedEntities) && resolvedEntities.length > 0)
-      ? `
-
-=== ENTIDADES RESUELTAS DEL CATÁLOGO (autoritativo, verificado en Apache AGE) ===
-El usuario mencionó las siguientes entidades. Para cada una, el catálogo Chagra confirma estos binomios CANÓNICOS. JAMÁS uses otro nombre científico, JAMÁS las confundas con géneros parecidos por sonido (gulupa NO es Psidium ni Cucurbita; aguacate NO es Psidium).
-
-${resolvedEntities.map((e) => `- "${e.mentioned}" (${e.kind}) → ${e.nombre_comun} = ${e.nombre_cientifico} [id: ${e.canonical_id}, confidence: ${e.confidence}]`).join('\n')}
-
-REGLA: si tu respuesta menciona cualquiera de estas entidades, USÁ el nombre científico EXACTO listado arriba. NO traduzcas, NO sustituyas, NO completes con otro género. Si dudas entre alternativas listadas, elige la de mayor confidence.
-=== FIN ENTIDADES RESUELTAS ===`
-      : '';
-
-    // NN2+NN3 bloque dinámico de análisis. Va al final del system prompt
-    // para que sea lo último que el LLM lee antes de la query — máxima
-    // proximidad. Le dice EXACTAMENTE qué tipo de query es y qué plagas
-    // canónicas usar.
-    const queryAnalysisBlock = `
-
-=== ANÁLISIS DE LA QUERY ACTUAL (frontend) ===
-- Tipo: ${analysis.topic}
-- Es enumerativa (CASO C aplica): ${analysis.isEnum ? 'SÍ — usa respuesta CASO C' : 'NO — IGNORA CASO C completamente, responde normal con tool evidence o conocimiento'}
-${analysis.pestsMentioned.length > 0 ? `- Plagas mencionadas (USA NOMBRE CIENTÍFICO EXACTO de abajo, NO inventes):
-${analysis.pestsMentioned.map((p) => `  · "${p.name}" → ${p.canonical}`).join('\n')}` : '- Plagas mencionadas: ninguna'}
-
-REGLA CRÍTICA SOBRE ESTE BLOQUE: este análisis es autoritativo para ESTA query. Si dice "Es enumerativa: NO", el CASO C del system prompt NO aplica aunque tu instinto te diga lo contrario. Si lista plagas, usa ESOS nombres científicos exactos (jamás otros, jamás "Fusarium spp" para chinches, jamás géneros inventados).
-=== FIN ANÁLISIS ===`;
-    // 2026-05-19: incidente alucinación tomate — el modelo confundía el
-    // corpus RAG con lo que el usuario dijo (atribuía síntomas "hojas
-    // amarillas" del documento de referencia al operador). Fix: delimitar
-    // EXPLÍCITAMENTE el corpus + instrucción literal de no citarlo como
-    // si fuera del usuario.
-    const corpusContext = contextCorpus.length > 0
-      ? `
-
-=== INFORMACIÓN DE REFERENCIA AGRONÓMICA (NO viene del usuario, NO citarla como si el usuario te lo hubiera contado) ===
-${contextCorpus.map((c) => c.text).join('\n\n---\n\n')}
-=== FIN REFERENCIA ===
-
-Usa esta referencia para informar tu respuesta, pero RESPONDE SOLO a lo que el usuario te preguntó. NO menciones síntomas ni observaciones que no estén explícitamente en el mensaje del usuario.`
-      : '';
+    // ENTIDADES RESUELTAS (DR taxonómico Tier 1 B) + análisis NN2/NN3 +
+    // corpus RAG delimitado — builders puros en agentPromptBase.
+    const resolvedEntitiesBlock = buildResolvedEntitiesBlock(resolvedEntities);
+    const queryAnalysisBlock = buildQueryAnalysisBlock(analysis);
+    const corpusContext = buildCorpusContext(contextCorpus);
 
     const evidenceContext = formatToolEvidence(toolEvidence);
